services/auth-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
nats-py==2.9.0
asyncpg==0.30.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
prometheus-client==0.21.1
httpx==0.28.1
